+++ MK4duo/src/core/commands/gcode/multimode/m3_m4.h
@@ -73,7 +73,6 @@
 
     } // printer.mode
 
-    mechanics.prepare_move_to_destination();
   }
 
   inline void gcode_M3(void) { gcode_M3_M4(true); }