+++ MK4duo/src/platform/HAL_DUE/HAL.h
@@ -283,7 +283,7 @@ class HAL {
     static bool pwm_status(const pin_t pin);
     static bool tc_status(const pin_t pin);
 
-    static void analogWrite(const pin_t pin, uint32_t ulValue, const uint16_t freq=1000);
+    static void analogWrite(const pin_t pin, const uint32_t value, const bool HWInvert=false, const uint16_t freq=1000);
 
     static void Tick();